ML/feature1/requirements.txt
google-genai
pytesseract
easyocr
python-docx
spacy
nltk
pydantic
pydantic-settings
numpy
pandas
Pillow
pytest
streamlit
fastapi
uvicorn[standard]

